requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via gradio
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.2
    # via realtime
aiosignal==1.3.2
    # via aiohttp
allure-pytest==2.14.2
    # via -r requirements.in
allure-python-commons==2.14.2
    # via allure-pytest
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.9.0
    # via
    #   -r requirements.in
    #   gradio
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   aiohttp
    #   allure-python-commons
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.4.26
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via
    #   -r requirements.in
    #   requests
click==8.2.0
    # via
    #   streamlit
    #   typer
    #   uvicorn
deprecated==1.2.18
    # via vecs
deprecation==2.1.0
    # via postgrest
distro==1.9.0
    # via
    #   -r requirements.in
    #   openai
    #   posthog
elevenlabs==1.59.0
    # via -r requirements.in
fastapi==0.115.12
    # via
    #   -r requirements.in
    #   gradio
ffmpy==0.5.0
    # via gradio
filelock==3.18.0
    # via huggingface-hub
flupy==1.2.2
    # via vecs
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via
    #   gradio-client
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
gotrue==2.12.0
    # via supabase
gradio==5.31.0
    # via -r requirements.in
gradio-client==1.10.1
    # via gradio
groovy==0.1.2
    # via gradio
grpcio==1.71.0
    # via qdrant-client
h11==0.16.0
    # via
    #   -r requirements.in
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hf-xet==1.1.2
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via
    #   -r requirements.in
    #   httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   elevenlabs
    #   gotrue
    #   gradio
    #   gradio-client
    #   openai
    #   postgrest
    #   qdrant-client
    #   safehttpx
    #   storage3
    #   supabase
    #   supafunc
huggingface-hub==0.32.2
    # via
    #   gradio
    #   gradio-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   gradio
    #   pydeck
jiter==0.9.0
    # via
    #   -r requirements.in
    #   openai
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mem0ai==0.1.102
    # via -r requirements.in
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
narwhals==1.41.0
    # via altair
numpy==2.2.6
    # via
    #   gradio
    #   pandas
    #   pgvector
    #   pydeck
    #   qdrant-client
    #   streamlit
openai==1.79.0
    # via
    #   -r requirements.in
    #   mem0ai
orjson==3.10.18
    # via gradio
packaging==24.2
    # via
    #   altair
    #   deprecation
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   pytest
    #   streamlit
pandas==2.2.3
    # via
    #   gradio
    #   streamlit
pgvector==0.3.6
    # via vecs
phonenumbers==9.0.6
    # via -r requirements.in
pillow==11.2.1
    # via
    #   -r requirements.in
    #   gradio
    #   streamlit
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
portalocker==2.10.1
    # via qdrant-client
postgrest==0.19.3
    # via supabase
posthog==4.2.0
    # via mem0ai
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.0
    # via
    #   qdrant-client
    #   streamlit
psycopg2-binary==2.9.10
    # via
    #   -r requirements.in
    #   vecs
pyarrow==20.0.0
    # via streamlit
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   elevenlabs
    #   fastapi
    #   gotrue
    #   gradio
    #   mem0ai
    #   openai
    #   postgrest
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via
    #   -r requirements.in
    #   elevenlabs
    #   pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pydeck==0.9.1
    # via streamlit
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via gotrue
pytest==8.3.5
    # via
    #   -r requirements.in
    #   allure-pytest
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.0.0
    # via -r requirements.in
pytest-mock==3.14.1
    # via gotrue
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   posthog
    #   realtime
    #   storage3
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via
    #   mem0ai
    #   pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
    #   uvicorn
qdrant-client==1.14.2
    # via
    #   -r requirements.in
    #   mem0ai
realtime==2.5.0
    # via supabase
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   elevenlabs
    #   huggingface-hub
    #   posthog
    #   streamlit
rich==14.0.0
    # via typer
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
ruff==0.11.11
    # via gradio
safehttpx==0.1.6
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   posthog
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   mem0ai
    #   vecs
starlette==0.46.2
    # via
    #   fastapi
    #   gradio
storage3==0.11.3
    # via supabase
streamlit==1.45.1
    # via -r requirements.in
strenum==0.4.15
    # via supafunc
supabase==2.13.0
    # via -r requirements.in
supafunc==0.9.4
    # via supabase
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tomlkit==0.13.2
    # via gradio
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
typer==0.16.0
    # via gradio
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   altair
    #   anyio
    #   elevenlabs
    #   fastapi
    #   flupy
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   realtime
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   -r requirements.in
    #   pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   -r requirements.in
    #   qdrant-client
    #   requests
uvicorn==0.34.2
    # via
    #   -r requirements.in
    #   gradio
uvloop==0.21.0
    # via uvicorn
vecs==0.4.5
    # via -r requirements.in
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   elevenlabs
    #   gradio-client
    #   realtime
    #   uvicorn
wrapt==1.17.2
    # via deprecated
yarl==1.20.0
    # via aiohttp
